pyproject.toml
[project]
name = "igdtuw-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.3.4",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.1",
    "google-genai>=1.46.0",
    "google-generativeai>=0.8.5",
    "gspread>=6.2.1",
    "hf-xet>=1.2.0",
    "huggingface-hub>=0.36.0",
    "oauth2client>=4.1.3",
    "pandas>=2.3.3",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "selenium>=4.36.0",
    "sentence-transformers>=5.1.2",
    "setuptools>=80.9.0",
    "tqdm>=4.67.1",
    "webdriver-manager>=4.0.2",
]
